+++ mod/app/src/main/java/bttv/api/Blacklist.java
@@ -3,7 +3,7 @@ package bttv.api;
 
 import android.util.Log;
 
-import tv.twitch.android.shared.chat.pub.ChatMessageInterface;
+import tv.twitch.android.shared.chat.pub.messages.ui.ChatMessageInterface;
 import tv.twitch.android.shared.chat.ChatMessageDelegate;
 
 public class Blacklist {
