+++ mod/app/src/main/java/bttv/api/Highlight.java
@@ -2,7 +2,7 @@ package bttv.api;
 
 import kotlin.jvm.functions.Function1;
 import android.content.Context;
-import tv.twitch.android.shared.chat.pub.ChatMessageInterface;
+import tv.twitch.android.shared.chat.pub.messages.ui.ChatMessageInterface;
 
 public class Highlight {
     public static Integer replaceNum(ChatMessageInterface chatMessageInterface, boolean z, boolean z2, Function1<Context, Integer> function, int i, Object iClickableUsernameSpanListener, Object twitchUrlSpanClickListener, Object webViewSource, String str, boolean z3, Object chatFiltersSettings, Integer num) {
