+++ mod/twitch/src/main/java/tv/twitch/android/shared/chat/pub/messages/ui/ChatMessageInterface.java
@@ -2,7 +2,7 @@
  * This file is a stub, so javac does not throw errors,
  * it is not used in the patches.
  */
-package tv.twitch.android.shared.chat.pub;
+package tv.twitch.android.shared.chat.pub.messages.ui;
 
 public interface ChatMessageInterface {
     java.util.List<tv.twitch.android.models.chat.MessageBadge> getBadges();
